Add time step preference & per session time steps
See: https://gitlab.nist.gov/gitlab/wnd-publicsafety/visualizer/NetSimulyzer/-/issues/114

See: https://gitlab.nist.gov/gitlab/wnd-publicsafety/visualizer/NetSimulyzer/-/issues/115

+++ src/window/scene/SceneWidget.cpp
@@ -496,7 +496,7 @@ void SceneWidget::setTime(double value) {
   emit timeChanged(simulationTime, diff);
 }
 
-void SceneWidget::setTimeStep(double value) {
+void SceneWidget::setTimeStep(int value) {
   timeStep = value;
 }
